Split examples into own test files
