+++ optional_test.go
@@ -22,30 +22,21 @@ package optional
 
 import (
 	"cmp"
-	"context"
 	"database/sql"
 	"database/sql/driver"
 	"encoding/json"
 	"encoding/xml"
 	"errors"
-	"fmt"
-	"github.com/neocotic/go-optional/internal/example"
 	"github.com/neocotic/go-optional/internal/test"
 	ptrs "github.com/neocotic/go-pointers"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/yaml.v3"
-	"log"
 	"math"
 	"strconv"
 	"strings"
 	"testing"
 	"time"
 	"unicode"
-)
-
-var (
-	ctx context.Context
-	db  *sql.DB
 )
 
 func BenchmarkOptional_Filter(b *testing.B) {
@@ -56,40 +47,6 @@ func BenchmarkOptional_Filter(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = opt.Filter(isPos)
 	}
-}
-
-func ExampleOptional_Filter_int() {
-	isPos := func(value int) bool {
-		return value >= 0
-	}
-
-	example.Print(Empty[int]().Filter(isPos))
-	example.Print(Of(-123).Filter(isPos))
-	example.Print(Of(0).Filter(isPos))
-	example.Print(Of(123).Filter(isPos))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// 0
-	// 123
-}
-
-func ExampleOptional_Filter_string() {
-	isLower := func(value string) bool {
-		return !strings.ContainsFunc(value, unicode.IsUpper)
-	}
-
-	example.Print(Empty[string]().Filter(isLower))
-	example.Print(Of("ABC").Filter(isLower))
-	example.Print(Of("").Filter(isLower))
-	example.Print(Of("abc").Filter(isLower))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// ""
-	// "abc"
 }
 
 type optionalFilterTC[T any] struct {
@@ -165,28 +122,6 @@ func BenchmarkOptional_Get(b *testing.B) {
 	}
 }
 
-func ExampleOptional_Get_int() {
-	example.PrintGet(Empty[int]().Get())
-	example.PrintGet(Of(0).Get())
-	example.PrintGet(Of(123).Get())
-
-	// Output:
-	// 0 false
-	// 0 true
-	// 123 true
-}
-
-func ExampleOptional_Get_string() {
-	example.PrintGet(Empty[string]().Get())
-	example.PrintGet(Of("").Get())
-	example.PrintGet(Of("abc").Get())
-
-	// Output:
-	// "" false
-	// "" true
-	// "abc" true
-}
-
 type optionalGetTC[T any] struct {
 	opt           Optional[T]
 	expectPresent bool
@@ -244,26 +179,6 @@ func BenchmarkOptional_IfPresent(b *testing.B) {
 	}
 }
 
-func ExampleOptional_IfPresent_int() {
-	Empty[int]().IfPresent(example.PrintValue[int]) // Does nothing
-	Of(0).IfPresent(example.PrintValue[int])
-	Of(123).IfPresent(example.PrintValue[int])
-
-	// Output:
-	// 0
-	// 123
-}
-
-func ExampleOptional_IfPresent_string() {
-	Empty[string]().IfPresent(example.PrintValue[string]) // Does nothing
-	Of("").IfPresent(example.PrintValue[string])
-	Of("abc").IfPresent(example.PrintValue[string])
-
-	// Output:
-	// ""
-	// "abc"
-}
-
 type optionalIfPresentTC[T any] struct {
 	opt             Optional[T]
 	expectCallCount uint
@@ -317,28 +232,6 @@ func BenchmarkOptional_IsEmpty(b *testing.B) {
 	}
 }
 
-func ExampleOptional_IsEmpty_int() {
-	fmt.Println(Empty[int]().IsEmpty())
-	fmt.Println(Of(0).IsEmpty())
-	fmt.Println(Of(123).IsEmpty())
-
-	// Output:
-	// true
-	// false
-	// false
-}
-
-func ExampleOptional_IsEmpty_string() {
-	fmt.Println(Empty[string]().IsEmpty())
-	fmt.Println(Of("").IsEmpty())
-	fmt.Println(Of("abc").IsEmpty())
-
-	// Output:
-	// true
-	// false
-	// false
-}
-
 type optionalIsEmptyTC[T any] struct {
 	opt    Optional[T]
 	expect bool
@@ -386,28 +279,6 @@ func BenchmarkOptional_IsPresent(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = opt.IsPresent()
 	}
-}
-
-func ExampleOptional_IsPresent_int() {
-	fmt.Println(Empty[int]().IsPresent())
-	fmt.Println(Of(0).IsPresent())
-	fmt.Println(Of(123).IsPresent())
-
-	// Output:
-	// false
-	// true
-	// true
-}
-
-func ExampleOptional_IsPresent_string() {
-	fmt.Println(Empty[string]().IsPresent())
-	fmt.Println(Of("").IsPresent())
-	fmt.Println(Of("abc").IsPresent())
-
-	// Output:
-	// false
-	// true
-	// true
 }
 
 type optionalIsPresentTC[T any] struct {
@@ -459,28 +330,6 @@ func BenchmarkOptional_IsZero(b *testing.B) {
 	}
 }
 
-func ExampleOptional_IsZero_int() {
-	fmt.Println(Empty[int]().IsZero())
-	fmt.Println(Of(0).IsZero())
-	fmt.Println(Of(123).IsZero())
-
-	// Output:
-	// true
-	// false
-	// false
-}
-
-func ExampleOptional_IsZero_string() {
-	fmt.Println(Empty[string]().IsZero())
-	fmt.Println(Of("").IsZero())
-	fmt.Println(Of("abc").IsZero())
-
-	// Output:
-	// true
-	// false
-	// false
-}
-
 type optionalIsZeroTC[T any] struct {
 	opt    Optional[T]
 	expect bool
@@ -530,39 +379,6 @@ func BenchmarkOptional_MarshalJSON(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_MarshalJSON() {
-	// json omitempty option does not apply to zero value structs
-	type MyStruct struct {
-		Number Optional[int]    `json:"number,omitempty"`
-		Text   Optional[string] `json:"text,omitempty"`
-	}
-
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{})))
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{Number: Of(0), Text: Of("")})))
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{Number: Of(123), Text: Of("abc")})))
-
-	// Output:
-	// {"number":null,"text":null} <nil>
-	// {"number":0,"text":""} <nil>
-	// {"number":123,"text":"abc"} <nil>
-}
-
-func ExampleOptional_MarshalJSON_pointers() {
-	type MyStruct struct {
-		Number *Optional[int]    `json:"number,omitempty"`
-		Text   *Optional[string] `json:"text,omitempty"`
-	}
-
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{})))
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{Number: ptrs.Value(Of(0)), Text: ptrs.Value(Of(""))})))
-	example.PrintMarshalled(json.Marshal(Of(MyStruct{Number: ptrs.Value(Of(123)), Text: ptrs.Value(Of("abc"))})))
-
-	// Output:
-	// {} <nil>
-	// {"number":0,"text":""} <nil>
-	// {"number":123,"text":"abc"} <nil>
 }
 
 type optionalMarshalJSONTC struct {
@@ -651,22 +467,6 @@ func BenchmarkOptional_MarshalXML(b *testing.B) {
 	}
 }
 
-func ExampleOptional_MarshalXML() {
-	type MyStruct struct {
-		Number Optional[int]    `xml:"number,omitempty"`
-		Text   Optional[string] `xml:"text,omitempty"`
-	}
-
-	example.PrintMarshalled(xml.Marshal(Of(MyStruct{})))
-	example.PrintMarshalled(xml.Marshal(Of(MyStruct{Number: Of(0), Text: Of("")})))
-	example.PrintMarshalled(xml.Marshal(Of(MyStruct{Number: Of(123), Text: Of("abc")})))
-
-	// Output:
-	// <MyStruct></MyStruct> <nil>
-	// <MyStruct><number>0</number><text></text></MyStruct> <nil>
-	// <MyStruct><number>123</number><text>abc</text></MyStruct> <nil>
-}
-
 type optionalMarshalXMLTC struct {
 	value     any
 	expectXML string
@@ -750,24 +550,6 @@ func BenchmarkOptional_MarshalYAML(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_MarshalYAML() {
-	type MyStruct struct {
-		Number Optional[int]    `yaml:"number,omitempty"`
-		Text   Optional[string] `yaml:"text,omitempty"`
-	}
-
-	example.PrintMarshalled(yaml.Marshal(Of(MyStruct{})))
-	example.PrintMarshalled(yaml.Marshal(Of(MyStruct{Number: Of(0), Text: Of("")})))
-	example.PrintMarshalled(yaml.Marshal(Of(MyStruct{Number: Of(123), Text: Of("abc")})))
-
-	// Output:
-	// {} <nil>
-	// number: 0
-	// text: "" <nil>
-	// number: 123
-	// text: abc <nil>
 }
 
 type optionalMarshalYAMLTC struct {
@@ -864,32 +646,6 @@ func BenchmarkOptional_OrElse(b *testing.B) {
 	}
 }
 
-func ExampleOptional_OrElse_int() {
-	defaultVal := -1
-
-	example.PrintValue(Empty[int]().OrElse(defaultVal))
-	example.PrintValue(Of(0).OrElse(defaultVal))
-	example.PrintValue(Of(123).OrElse(defaultVal))
-
-	// Output:
-	// -1
-	// 0
-	// 123
-}
-
-func ExampleOptional_OrElse_string() {
-	defaultVal := "unknown"
-
-	example.PrintValue(Empty[string]().OrElse(defaultVal))
-	example.PrintValue(Of("").OrElse(defaultVal))
-	example.PrintValue(Of("abc").OrElse(defaultVal))
-
-	// Output:
-	// "unknown"
-	// ""
-	// "abc"
-}
-
 type optionalOrElseTC[T any] struct {
 	opt    Optional[T]
 	other  T
@@ -949,36 +705,6 @@ func BenchmarkOptional_OrElseGet(b *testing.B) {
 			return -1
 		})
 	}
-}
-
-func ExampleOptional_OrElseGet_int() {
-	defaultFunc := func() int {
-		return -1
-	}
-
-	example.PrintValue(Empty[int]().OrElseGet(defaultFunc))
-	example.PrintValue(Of(0).OrElseGet(defaultFunc))
-	example.PrintValue(Of(123).OrElseGet(defaultFunc))
-
-	// Output:
-	// -1
-	// 0
-	// 123
-}
-
-func ExampleOptional_OrElseGet_string() {
-	defaultFunc := func() string {
-		return "unknown"
-	}
-
-	example.PrintValue(Empty[string]().OrElseGet(defaultFunc))
-	example.PrintValue(Of("").OrElseGet(defaultFunc))
-	example.PrintValue(Of("abc").OrElseGet(defaultFunc))
-
-	// Output:
-	// "unknown"
-	// ""
-	// "abc"
 }
 
 type optionalOrElseGetTC[T any] struct {
@@ -1049,43 +775,6 @@ func BenchmarkOptional_OrElseTryGet(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_OrElseTryGet_int() {
-	defaultFunc := func() (int, error) {
-		return -1, nil
-	}
-
-	example.PrintTryValue(Empty[int]().OrElseTryGet(defaultFunc))
-	example.PrintTryValue(Of(0).OrElseTryGet(defaultFunc))
-	example.PrintTryValue(Of(123).OrElseTryGet(defaultFunc))
-
-	// Output:
-	// -1 <nil>
-	// 0 <nil>
-	// 123 <nil>
-}
-
-func ExampleOptional_OrElseTryGet_string() {
-	var defaultStringUsed bool
-	defaultFunc := func() (string, error) {
-		if defaultStringUsed {
-			return "", errors.New("default string already used")
-		}
-		defaultStringUsed = true
-		return "unknown", nil
-	}
-
-	example.PrintTryValue(Empty[string]().OrElseTryGet(defaultFunc))
-	example.PrintTryValue(Of("").OrElseTryGet(defaultFunc))
-	example.PrintTryValue(Of("abc").OrElseTryGet(defaultFunc))
-	example.PrintTryValue(Empty[string]().OrElseTryGet(defaultFunc))
-
-	// Output:
-	// "unknown" <nil>
-	// "" <nil>
-	// "abc" <nil>
-	// "" "default string already used"
 }
 
 type optionalOrElseTryGetTC[T any] struct {
@@ -1169,34 +858,6 @@ func BenchmarkOptional_Require(b *testing.B) {
 	}
 }
 
-func ExampleOptional_Require_int() {
-	example.PrintValue(Of(0).Require())
-	example.PrintValue(Of(123).Require())
-
-	// Output:
-	// 0
-	// 123
-}
-
-func ExampleOptional_Require_panic() {
-	defer func() {
-		fmt.Println(recover())
-	}()
-
-	Empty[int]().Require()
-
-	// Output: go-optional: value not present
-}
-
-func ExampleOptional_Require_string() {
-	example.PrintValue(Of("").Require())
-	example.PrintValue(Of("abc").Require())
-
-	// Output:
-	// ""
-	// "abc"
-}
-
 type optionalRequireTC[T any] struct {
 	opt         Optional[T]
 	expectPanic bool
@@ -1256,31 +917,6 @@ func BenchmarkOptional_Scan(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_Scan() {
-	rows, err := db.QueryContext(ctx, "SELECT name, age FROM users")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer rows.Close()
-
-	users := make(map[string]Optional[int])
-	for rows.Next() {
-		var (
-			age  Optional[int]
-			name string
-		)
-		if err = rows.Scan(&name, &age); err != nil {
-			log.Fatal(err)
-		}
-		users[name] = age
-	}
-	if err = rows.Err(); err != nil {
-		log.Fatal(err)
-	}
-
-	log.Printf("user demographics: %s", users)
 }
 
 type optionalScanTC[S, T any] struct {
@@ -4725,28 +4361,6 @@ func BenchmarkOptional_String(b *testing.B) {
 	}
 }
 
-func ExampleOptional_String_int() {
-	fmt.Printf("%q\n", Empty[int]().String())
-	fmt.Printf("%q\n", Of(0).String())
-	fmt.Printf("%q\n", Of(123).String())
-
-	// Output:
-	// "<empty>"
-	// "0"
-	// "123"
-}
-
-func ExampleOptional_String_string() {
-	fmt.Printf("%q\n", Empty[string]().String())
-	fmt.Printf("%q\n", Of("").String())
-	fmt.Printf("%q\n", Of("abc").String())
-
-	// Output:
-	// "<empty>"
-	// ""
-	// "abc"
-}
-
 type optionalStringTC[T any] struct {
 	opt    Optional[T]
 	expect string
@@ -4796,40 +4410,6 @@ func BenchmarkOptional_UnmarshalJSON(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_UnmarshalJSON() {
-	type MyStruct struct {
-		Number Optional[int]    `json:"number"`
-		Text   Optional[string] `json:"text"`
-	}
-
-	inputs := []string{
-		`{}`,
-		`{"number":null,"text":null}`,
-		`{"number":0,"text":""}`,
-		`{"number":123,"text":"abc"}`,
-	}
-
-	for _, input := range inputs {
-		var output MyStruct
-		if err := json.Unmarshal([]byte(input), &output); err != nil {
-			log.Fatal(err)
-		}
-
-		example.Print(output.Number)
-		example.Print(output.Text)
-	}
-
-	// Output:
-	// <empty>
-	// <empty>
-	// 0
-	// ""
-	// 0
-	// ""
-	// 123
-	// "abc"
 }
 
 type optionalUnmarshalJSONTC[T any] struct {
@@ -4912,37 +4492,6 @@ func BenchmarkOptional_UnmarshalXML(b *testing.B) {
 	}
 }
 
-func ExampleOptional_UnmarshalXML() {
-	type MyStruct struct {
-		Number Optional[int]    `xml:"number"`
-		Text   Optional[string] `xml:"text"`
-	}
-
-	inputs := []string{
-		`<MyStruct></MyStruct>`,
-		`<MyStruct><number>0</number><text></text></MyStruct>`,
-		`<MyStruct><number>123</number><text>abc</text></MyStruct>`,
-	}
-
-	for _, input := range inputs {
-		var output MyStruct
-		if err := xml.Unmarshal([]byte(input), &output); err != nil {
-			log.Fatal(err)
-		}
-
-		example.Print(output.Number)
-		example.Print(output.Text)
-	}
-
-	// Output:
-	// <empty>
-	// <empty>
-	// 0
-	// ""
-	// 123
-	// "abc"
-}
-
 type optionalUnmarshalXMLTC[T any] struct {
 	xml    string
 	expect T
@@ -5021,43 +4570,6 @@ func BenchmarkOptional_UnmarshalYAML(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleOptional_UnmarshalYAML() {
-	type MyStruct struct {
-		Number Optional[int]    `json:"number"`
-		Text   Optional[string] `json:"text"`
-	}
-
-	inputs := []string{
-		`{}`,
-		`number: null
-text: null`,
-		`number: 0
-text: ""`,
-		`number: 123
-text: abc`,
-	}
-
-	for _, input := range inputs {
-		var output MyStruct
-		if err := yaml.Unmarshal([]byte(input), &output); err != nil {
-			log.Fatal(err)
-		}
-
-		example.Print(output.Number)
-		example.Print(output.Text)
-	}
-
-	// Output:
-	// <empty>
-	// <empty>
-	// <empty>
-	// <empty>
-	// 0
-	// ""
-	// 123
-	// "abc"
 }
 
 type optionalUnmarshalYAMLTC[T any] struct {
@@ -5143,22 +4655,6 @@ func BenchmarkOptional_Value(b *testing.B) {
 		if _, err := opt.Value(); err != nil {
 			b.Fatal(err)
 		}
-	}
-}
-
-func ExampleOptional_Value() {
-	username := "alex"
-	age := Of(30)
-	result, err := db.ExecContext(ctx, "UPDATE users SET age = ? WHERE username = ?", age, username)
-	if err != nil {
-		log.Fatal(err)
-	}
-	rows, err := result.RowsAffected()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if rows != 1 {
-		log.Fatalf("expected to affect 1 row, affected %d", rows)
 	}
 }
 
@@ -5327,48 +4823,6 @@ func BenchmarkCompare(b *testing.B) {
 	}
 }
 
-func ExampleCompare_int() {
-	fmt.Println(Compare(Empty[int](), Of(0)))
-	fmt.Println(Compare(Of(0), Of(123)))
-
-	fmt.Println(Compare(Empty[int](), Empty[int]()))
-	fmt.Println(Compare(Of(0), Of(0)))
-	fmt.Println(Compare(Of(123), Of(123)))
-
-	fmt.Println(Compare(Of(0), Empty[int]()))
-	fmt.Println(Compare(Of(123), Of(0)))
-
-	// Output:
-	// -1
-	// -1
-	// 0
-	// 0
-	// 0
-	// 1
-	// 1
-}
-
-func ExampleCompare_string() {
-	fmt.Println(Compare(Empty[string](), Of("")))
-	fmt.Println(Compare(Of(""), Of("abc")))
-
-	fmt.Println(Compare(Empty[string](), Empty[string]()))
-	fmt.Println(Compare(Of(""), Of("")))
-	fmt.Println(Compare(Of("abc"), Of("abc")))
-
-	fmt.Println(Compare(Of(""), Empty[string]()))
-	fmt.Println(Compare(Of("abc"), Of("")))
-
-	// Output:
-	// -1
-	// -1
-	// 0
-	// 0
-	// 0
-	// 1
-	// 1
-}
-
 type compareTC[T cmp.Ordered] struct {
 	x      Optional[T]
 	y      Optional[T]
@@ -5429,18 +4883,6 @@ func BenchmarkEmpty(b *testing.B) {
 	}
 }
 
-func ExampleEmpty_int() {
-	example.Print(Empty[int]())
-
-	// Output: <empty>
-}
-
-func ExampleEmpty_string() {
-	example.Print(Empty[string]())
-
-	// Output: <empty>
-}
-
 type emptyTC[T any] struct {
 	test.Control
 }
@@ -5466,32 +4908,6 @@ func BenchmarkFind(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = Find(opts...)
 	}
-}
-
-func ExampleFind_int() {
-	example.Print(Find[int]())
-	example.Print(Find(Empty[int]()))
-	example.Print(Find(Empty[int](), Of(0), Of(123)))
-	example.Print(Find(Empty[int](), Of(123), Of(0)))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// 0
-	// 123
-}
-
-func ExampleFind_string() {
-	example.Print(Find[string]())
-	example.Print(Find(Empty[string]()))
-	example.Print(Find(Empty[string](), Of(""), Of("abc")))
-	example.Print(Find(Empty[string](), Of("abc"), Of("")))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// ""
-	// "abc"
 }
 
 type findTC[T any] struct {
@@ -5562,48 +4978,6 @@ func BenchmarkFlatMap(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = FlatMap(opt, toString)
 	}
-}
-
-func ExampleFlatMap_int() {
-	mapper := func(value int) Optional[string] {
-		if value == 0 {
-			return Empty[string]()
-		}
-		return Of(strconv.FormatInt(int64(value), 10))
-	}
-
-	example.Print(FlatMap(Empty[int](), mapper))
-	example.Print(FlatMap(Of(0), mapper))
-	example.Print(FlatMap(Of(123), mapper))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// "123"
-}
-
-func ExampleFlatMap_string() {
-	mapper := func(value string) Optional[int] {
-		if value == "" {
-			return Empty[int]()
-		}
-		i, err := strconv.ParseInt(value, 10, 0)
-		if err != nil {
-			log.Fatal(err)
-		}
-		return OfZeroable(int(i))
-	}
-
-	example.Print(FlatMap(Empty[string](), mapper))
-	example.Print(FlatMap(Of(""), mapper))
-	example.Print(FlatMap(Of("0"), mapper))
-	example.Print(FlatMap(Of("123"), mapper))
-
-	// Output:
-	// <empty>
-	// <empty>
-	// <empty>
-	// 123
 }
 
 type flatMapTC[T, M any] struct {
@@ -5689,28 +5063,6 @@ func BenchmarkGetAny(b *testing.B) {
 	}
 }
 
-func ExampleGetAny_int() {
-	example.PrintValues(GetAny[int]())
-	example.PrintValues(GetAny(Empty[int]()))
-	example.PrintValues(GetAny(Empty[int](), Of(0), Of(123)))
-
-	// Output:
-	// []
-	// []
-	// [0 123]
-}
-
-func ExampleGetAny_string() {
-	example.PrintValues(GetAny[string]())
-	example.PrintValues(GetAny(Empty[string]()))
-	example.PrintValues(GetAny(Empty[string](), Of("abc"), Of("")))
-
-	// Output:
-	// []
-	// []
-	// ["abc" ""]
-}
-
 type getAnyTC[T any] struct {
 	opts   []Optional[T]
 	expect []T
@@ -5767,40 +5119,6 @@ func BenchmarkMap(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = Map(opt, toString)
 	}
-}
-
-func ExampleMap_int() {
-	mapper := func(value int) string {
-		return strconv.FormatInt(int64(value), 10)
-	}
-
-	example.Print(Map(Empty[int](), mapper))
-	example.Print(Map(Of(0), mapper))
-	example.Print(Map(Of(123), mapper))
-
-	// Output:
-	// <empty>
-	// "0"
-	// "123"
-}
-
-func ExampleMap_string() {
-	mapper := func(value string) int {
-		i, err := strconv.ParseInt(value, 10, 0)
-		if err != nil {
-			log.Fatal(err)
-		}
-		return int(i)
-	}
-
-	example.Print(Map(Empty[string](), mapper))
-	example.Print(Map(Of("0"), mapper))
-	example.Print(Map(Of("123"), mapper))
-
-	// Output:
-	// <empty>
-	// 0
-	// 123
 }
 
 type mapTC[T, M any] struct {
@@ -5877,28 +5195,6 @@ func BenchmarkMustFind(b *testing.B) {
 	}
 }
 
-func ExampleMustFind_int() {
-	example.PrintValue(MustFind(Empty[int](), Of(0), Of(123)))
-
-	// Output: 0
-}
-
-func ExampleMustFind_panic() {
-	defer func() {
-		fmt.Println(recover())
-	}()
-
-	MustFind(Empty[int]())
-
-	// Output: go-optional: value not present
-}
-
-func ExampleMustFind_string() {
-	example.PrintValue(MustFind(Empty[string](), Of("abc"), Of("")))
-
-	// Output: "abc"
-}
-
 type mustFindTC[T any] struct {
 	opts        []Optional[T]
 	expectPanic bool
@@ -5963,46 +5259,6 @@ func BenchmarkOf(b *testing.B) {
 	}
 }
 
-func ExampleOf_int() {
-	example.Print(Of(0))
-	example.Print(Of(123))
-
-	// Output:
-	// 0
-	// 123
-}
-
-func ExampleOf_int_pointer() {
-	example.Print(Of((*int)(nil)))
-	example.Print(Of(ptrs.ZeroInt()))
-	example.Print(Of(ptrs.Int(123)))
-
-	// Output:
-	// <nil>
-	// &0
-	// &123
-}
-
-func ExampleOf_string() {
-	example.Print(Of(""))
-	example.Print(Of("abc"))
-
-	// Output:
-	// ""
-	// "abc"
-}
-
-func ExampleOf_string_pointer() {
-	example.Print(Of((*string)(nil)))
-	example.Print(Of(ptrs.ZeroString()))
-	example.Print(Of(ptrs.String("abc")))
-
-	// Output:
-	// <nil>
-	// &""
-	// &"abc"
-}
-
 type ofTC[T any] struct {
 	value T
 	test.Control
@@ -6057,46 +5313,6 @@ func BenchmarkOfNillable(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = OfNillable(&value)
 	}
-}
-
-func ExampleOfNillable_int() {
-	example.Print(OfNillable(0))
-	example.Print(OfNillable(123))
-
-	// Output:
-	// 0
-	// 123
-}
-
-func ExampleOfNillable_int_pointer() {
-	example.Print(OfNillable((*int)(nil)))
-	example.Print(OfNillable(ptrs.ZeroInt()))
-	example.Print(OfNillable(ptrs.Int(123)))
-
-	// Output:
-	// <empty>
-	// &0
-	// &123
-}
-
-func ExampleOfNillable_string() {
-	example.Print(OfNillable(""))
-	example.Print(OfNillable("abc"))
-
-	// Output:
-	// ""
-	// "abc"
-}
-
-func ExampleOfNillable_string_pointer() {
-	example.Print(OfNillable((*string)(nil)))
-	example.Print(OfNillable(ptrs.ZeroString()))
-	example.Print(OfNillable(ptrs.String("abc")))
-
-	// Output:
-	// <empty>
-	// &""
-	// &"abc"
 }
 
 type ofNillableTC[T any] struct {
@@ -6165,24 +5381,6 @@ func BenchmarkOfPointer(b *testing.B) {
 	}
 }
 
-func ExampleOfPointer_int() {
-	example.Print(OfPointer(0))
-	example.Print(OfPointer(123))
-
-	// Output:
-	// &0
-	// &123
-}
-
-func ExampleOfPointer_string() {
-	example.Print(OfPointer(""))
-	example.Print(OfPointer("abc"))
-
-	// Output:
-	// &""
-	// &"abc"
-}
-
 type ofPointerTC[T any] struct {
 	value T
 	test.Control
@@ -6219,46 +5417,6 @@ func BenchmarkOfZeroable(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = OfZeroable(123)
 	}
-}
-
-func ExampleOfZeroable_int() {
-	example.Print(OfZeroable(0))
-	example.Print(OfZeroable(123))
-
-	// Output:
-	// <empty>
-	// 123
-}
-
-func ExampleOfZeroable_int_pointer() {
-	example.Print(OfZeroable((*int)(nil)))
-	example.Print(OfZeroable(ptrs.ZeroInt()))
-	example.Print(OfZeroable(ptrs.Int(123)))
-
-	// Output:
-	// <empty>
-	// &0
-	// &123
-}
-
-func ExampleOfZeroable_string() {
-	example.Print(OfZeroable(""))
-	example.Print(OfZeroable("abc"))
-
-	// Output:
-	// <empty>
-	// "abc"
-}
-
-func ExampleOfZeroable_string_pointer() {
-	example.Print(OfZeroable((*string)(nil)))
-	example.Print(OfZeroable(ptrs.ZeroString()))
-	example.Print(OfZeroable(ptrs.String("abc")))
-
-	// Output:
-	// <empty>
-	// &""
-	// &"abc"
 }
 
 type ofZeroableTC[T any] struct {
@@ -6326,28 +5484,6 @@ func BenchmarkRequireAny(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		_ = RequireAny(opts...)
 	}
-}
-
-func ExampleRequireAny_int() {
-	example.PrintValues(RequireAny(Empty[int](), Of(0), Of(123)))
-
-	// Output: [0 123]
-}
-
-func ExampleRequireAny_panic() {
-	defer func() {
-		fmt.Println(recover())
-	}()
-
-	RequireAny(Empty[int]())
-
-	// Output: go-optional: value not present
-}
-
-func ExampleRequireAny_string() {
-	example.PrintValues(RequireAny(Empty[string](), Of(""), Of("abc")))
-
-	// Output: ["" "abc"]
 }
 
 type requireAnyTC[T any] struct {
@@ -6421,50 +5557,6 @@ func BenchmarkTryFlatMap(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleTryFlatMap_int() {
-	mapper := func(value int) (Optional[string], error) {
-		if value == 0 {
-			return Empty[string](), nil
-		}
-		return Of(strconv.FormatInt(int64(value), 10)), nil
-	}
-
-	example.PrintTry(TryFlatMap(Empty[int](), mapper))
-	example.PrintTry(TryFlatMap(Of(0), mapper))
-	example.PrintTry(TryFlatMap(Of(123), mapper))
-
-	// Output:
-	// <empty> <nil>
-	// <empty> <nil>
-	// "123" <nil>
-}
-
-func ExampleTryFlatMap_string() {
-	mapper := func(value string) (Optional[int], error) {
-		if value == "" {
-			return Empty[int](), nil
-		}
-		i, err := strconv.ParseInt(value, 10, 0)
-		if err != nil {
-			return Empty[int](), err
-		}
-		return OfZeroable(int(i)), nil
-	}
-
-	example.PrintTry(TryFlatMap(Empty[string](), mapper))
-	example.PrintTry(TryFlatMap(Of(""), mapper))
-	example.PrintTry(TryFlatMap(Of("0"), mapper))
-	example.PrintTry(TryFlatMap(Of("123"), mapper))
-	example.PrintTry(TryFlatMap(Of("abc"), mapper))
-
-	// Output:
-	// <empty> <nil>
-	// <empty> <nil>
-	// <empty> <nil>
-	// 123 <nil>
-	// <empty> "strconv.ParseInt: parsing \"abc\": invalid syntax"
 }
 
 type tryFatMapTC[T, M any] struct {
@@ -6564,39 +5656,6 @@ func BenchmarkTryMap(b *testing.B) {
 			b.Fatal(err)
 		}
 	}
-}
-
-func ExampleTryMap_int() {
-	mapper := func(value int) (string, error) {
-		return strconv.FormatInt(int64(value), 10), nil
-	}
-
-	example.PrintTry(TryMap(Empty[int](), mapper))
-	example.PrintTry(TryMap(Of(0), mapper))
-	example.PrintTry(TryMap(Of(123), mapper))
-
-	// Output:
-	// <empty> <nil>
-	// "0" <nil>
-	// "123" <nil>
-}
-
-func ExampleTryMap_string() {
-	mapper := func(value string) (int, error) {
-		i, err := strconv.ParseInt(value, 10, 0)
-		return int(i), err
-	}
-
-	example.PrintTry(TryMap(Empty[string](), mapper))
-	example.PrintTry(TryMap(Of("0"), mapper))
-	example.PrintTry(TryMap(Of("123"), mapper))
-	example.PrintTry(TryMap(Of("abc"), mapper))
-
-	// Output:
-	// <empty> <nil>
-	// 0 <nil>
-	// 123 <nil>
-	// <empty> "strconv.ParseInt: parsing \"abc\": invalid syntax"
 }
 
 type tryMapTC[T, M any] struct {
